+++ src/types/runtime/index.ts
@@ -2,13 +2,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import type { CellEvent, NotebookContent, CellTypeDefinition, CellPropertyDefinition, ControlsDefinition, IconTemplate, Cell } from "../types";
-import type { ConsoleCatcher } from "../console/console";
-import type { CellElement } from "../components/cell";
-import type { StarboardNotebookElement } from "../components/notebook";
+import type { CellEvent, NotebookContent, CellTypeDefinition, CellPropertyDefinition, ControlsDefinition, IconTemplate, Cell } from "..";
+import type { ConsoleCatcher } from "../../console/console";
+import type { CellElement } from "../../components/cell";
+import type { StarboardNotebookElement } from "../../components/notebook";
 import type { TemplateResult } from "lit-html";
-import type { StarboardTextEditor } from "../components/textEditor";
-import type { ConsoleOutputElement } from "../components/output/consoleOutput";
+import type { StarboardTextEditor } from "../../components/textEditor";
+import type { ConsoleOutputElement } from "../../components/output/consoleOutput";
 
 import type * as lithtmlLibrary from "lit-html";
 import type * as litElementLibrary from "lit-element";
@@ -17,19 +17,17 @@ import type * as YAML from "yaml";
 import type mdlib from "markdown-it";
 import type * as Popper from "@popperjs/core";
 
-import type { JavascriptEvaluator } from "../cellTypes/javascript/eval";
-import type { hookMarkdownItToPrismHighlighter } from "../components/helpers/highlight";
-import type { createCellProxy } from "../components/helpers/cellProxy";
-import type { cellToText, notebookContentToText } from "../content/serialization";
-import type { precompileJavascriptCode } from "../cellTypes/javascript/precompile";
-import type { MapRegistry } from "./registry";
-import type { hookMarkdownItToKaTeX } from "../components/helpers/katex";
-import type { renderIfHtmlOutput } from "../components/output/htmlOutput";
-import type { hookMarkdownItToEmojiPlugin } from "../components/helpers/emoji";
-import type { OutboundNotebookMessage } from "../messages/types";
-import type { StarboardContentEditor } from "../components/editor/contentEditor";
-
-export * from "../types";
+import type { JavascriptEvaluator } from "../../cellTypes/javascript/eval";
+import type { hookMarkdownItToPrismHighlighter } from "../../components/helpers/highlight";
+import type { createCellProxy } from "../../components/helpers/cellProxy";
+import type { cellToText, notebookContentToText } from "../../content/serialization";
+import type { precompileJavascriptCode } from "../../cellTypes/javascript/precompile";
+import type { MapRegistry } from "./../registry";
+import type { hookMarkdownItToKaTeX } from "../../components/helpers/katex";
+import type { renderIfHtmlOutput } from "../../components/output/htmlOutput";
+import type { hookMarkdownItToEmojiPlugin } from "../../components/helpers/emoji";
+import type { OutboundNotebookMessage } from "../messages";
+import type { StarboardContentEditor } from "../../components/editor/contentEditor";
 
 export interface RuntimeControls {
     insertCell(data: Partial<Cell>, position: "end" | "before" | "after", adjacentCellId?: string): void;
@@ -37,7 +35,7 @@ export interface RuntimeControls {
     changeCellType(id: string, newCellType: string): void;
     resetCell(id: string): void;
     runCell(id: string, focusNext?: boolean, insertNewCell?: boolean): void;
-    runAllCells(opts: {onlyRunOnLoad?: boolean}): Promise<void>;
+    runAllCells(opts: { onlyRunOnLoad?: boolean }): Promise<void>;
 
     /** 
      * Requests a save operation from the parent iframe.
@@ -65,8 +63,8 @@ export interface RuntimeControls {
      * @param id 
      * @param callback 
      */
-    subscribeToCellChanges(id: string, callback: () => void):  void;
-    unsubscribeToCellChanges(id: string, callback: () => void):  void;
+    subscribeToCellChanges(id: string, callback: () => void): void;
+    unsubscribeToCellChanges(id: string, callback: () => void): void;
 }
 
 
@@ -142,7 +140,7 @@ export interface Runtime {
      * The state of the notebook that exactly describes the text in the notebook.
      */
     content: NotebookContent;
- 
+
     definitions: {
         /**
          * Map of registered cell types, indexed by cellType (e.g. "javascript").
@@ -172,7 +170,7 @@ export interface Runtime {
      * Version of Starboard Notebook
      */
     version: string;
-    
+
     /**
      * Name of the runtime.
      */
@@ -195,10 +193,10 @@ export interface Runtime {
      */
     internal: {
         listeners: {
-            cellContentChanges: Map<string, (()=>void)[]>;
+            cellContentChanges: Map<string, (() => void)[]>;
         };
     };
-    
+
 }
 
 /**